Plasma clinical analytes

A set of 9 common clinical analytes measured in plasma.

ANALYTES

Format

An object of class data.frame with 146 rows and 20 columns.

plate

integer; 96-well plate identifier (1–4).

runseq

integer; run order (1–146).

platepos

integer; position in each 96-well plate.

pid

integer; randomly generated 8-digit identifier used in linkage to phenotypic data. All samples from the same animal have the same PID.

bid

integer; unique 5 digit identifier of all samples collected for an acute test/sample collection period. All samples collected during that period will have the same BID.

viallabel

integer; character, unique 11 digit numeric identifier of a sample vial.

omics_subset

logical; whether the sample was selected for -omics analysis.

sex

factor; the sex of the rat with levels "Female" and "Male".

timepoint

factor; exercise training group. Either "SED" (sedentary) or the number of weeks of training ("1W", "2W", "4W", "8W").

corticosterone

numeric; corticosterone (ng/mL).

glucagon

numeric; glucagon (pmol).

glucose

numeric; glucose (mg/dL).

glycerol

numeric; glycerol (mg/dL).

insulin

numeric; insulin (pg/mL).

insulin_iu

numeric; insulin (\(\mu\)IU/mL) using the conversion 1 pg/mL = 0.023 \(\mu\)IU/mL provided in the package insert.

insulin_pm

numeric; insulin (pM) calculated by dividing insulin (pg/mL) by its molecular weight (5.804 kDa).

lactate

numeric; lactate (mmol/L).

leptin

numeric; leptin (pg/mL).

nefa

numeric; non-esterified fatty acids (NEFA) (mmol/L).

total_ketones

numeric; total ketone bodies (\(\mu\)mol/L).

Details

Analyte	Catalogue No.	Company

Corticosterone	55-CORMS-E01	Alpco (Salem, NH)

Glucose	B24985	Beckman Coulter (Brea, CA)

Glycerol	445850	Beckman Coulter

Lactate	A95550	Beckman Coulter

Total Ketones	415-73301, 411-73401	Fujifilm Wako (Osaka, Japan)

NEFA	995-34791, 999-34691, 993-35191, 991-34891	Fujifilm Wako

Glucagon	K1535YK	Meso Scale Discovery (Rockville, MD)

Insulin, Leptin	K15158C (multiplex assay)	Meso Scale Discovery

See also