Metabolomics RefMet chemical subclass FGSEA

FGSEA of the metabolomics differential analysis results using the Metabolomics Workbench RefMet chemical subclasses.

METAB_FGSEA

Format

A named list of 3 data.frame objects. Similar to what is in GO_FGSEA.

Details

Metabolites were grouped according to their RefMet subclass and FGSEA was performed the same as for gene sets. Only those subclasses with at least 10 metabolites, after filtering to what was detected in WAT, were tested for enrichment.

References

Fahy, E., & Subramaniam, S. (2020). RefMet: A reference nomenclature for metabolomics. Nature Methods, 17(12), 1173–1174. https://doi.org/10.1038/s41592-020-01009-y

See also