Metabolomics MSnSet

An object of class MSnbase::MSnSet containing log\(_2\)-transformed metabolite data for differential analysis and WGCNA.

Usage

METAB_MSNSET

Format

Object of class MSnbase::MSnSet with 1063 features and 50 samples.

assayData

Object of class matrix with 1063 rows and 50 columns. See METAB_NORM_DATA_NESTED for more details.

featureData

Object of class data.frame with 1063 rows and 13 variables containing additional feature information:

feature_ID

character; metabolite RefMet identifier.

dataset

character; metabolomics platform in which the feature was measured.

name_in_figures

character; alternative feature identifier primarily used in visualizations.

refmet_super_class

character; RefMet chemical super class.

refmet_main_class

character; RefMet chemical main class.

refmet_sub_class

character; RefMet chemical sub class.

lipid_class

character; (lipids only) the class of the lipid. There are 43 unique classes.

chain_length

integer; (lipids only) total number of carbons that comprise the fatty acid chains.

double_bond

integer; total number of C=C double bonds in the fatty acid chains.

rt

numeric; retention time. Used to assign lipid_class.

mz

numeric; mass-to-charge ratio (m/z).

neutral_mass

numeric; neutral mass (g/mol).

formula

character; chemical formula.

phenoData

Object of class data.frame with 50 rows and 6 variables containing additional sample information:

pid

integer; randomly generated 8-digit identifier used in linkage to phenotypic data. All samples from the same animal have the same PID.

bid

integer; unique 5 digit identifier of all samples collected for an acute test/sample collection period. All samples collected during that period will have the same BID.

labelid

integer; unique 11 digit specimen label identifier, originating at the collection site, that provides a link to specimen processing and is used for shipments to the biorepository. Same as viallabel only in instances where aliquots are not further processed at the biorepository.

sex

factor; the sex of the rat with levels "Female" and "Male".

timepoint

factor; exercise training group. Either "SED" (sedentary) or the number of weeks of training ("1W", "2W", "4W", "8W").

exp_group

factor; experimental group. Unique combination of sex (first letter) and timepoint.

References

Fahy, E., & Subramaniam, S. (2020). RefMet: A reference nomenclature for metabolomics. Nature Methods, 17(12), 1173--1174. https://doi.org/10.1038/s41592-020-01009-y

See also

Examples

library(MSnbase)
#> Loading required package: BiocGenerics
#> 
#> Attaching package: ‘BiocGenerics’
#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from ‘package:base’:
#> 
#>     Filter, Find, Map, Position, Reduce, anyDuplicated, aperm, append,
#>     as.data.frame, basename, cbind, colnames, dirname, do.call,
#>     duplicated, eval, evalq, get, grep, grepl, intersect, is.unsorted,
#>     lapply, mapply, match, mget, order, paste, pmax, pmax.int, pmin,
#>     pmin.int, rank, rbind, rownames, sapply, setdiff, sort, table,
#>     tapply, union, unique, unsplit, which.max, which.min
#> Loading required package: Biobase
#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.
#> Loading required package: mzR
#> Loading required package: Rcpp
#> Loading required package: S4Vectors
#> Loading required package: stats4
#> 
#> Attaching package: ‘S4Vectors’
#> The following objects are masked from ‘package:base’:
#> 
#>     I, expand.grid, unname
#> Loading required package: ProtGenerics
#> 
#> Attaching package: ‘ProtGenerics’
#> The following object is masked from ‘package:stats’:
#> 
#>     smooth
#> 
#> This is MSnbase version 2.24.2 
#>   Visit https://lgatto.github.io/MSnbase/ to get started.
#> 
#> Attaching package: ‘MSnbase’
#> The following object is masked from ‘package:base’:
#> 
#>     trimws
METAB_MSNSET                   # summary
#> MSnSet (storageMode: lockedEnvironment)
#> assayData: 1063 features, 50 samples 
#>   element names: exprs 
#> protocolData: none
#> phenoData
#>   sampleNames: 90245014701 90248014701 ... 90227014701 (50 total)
#>   varLabels: pid bid ... exp_group (6 total)
#>   varMetadata: labelDescription
#> featureData
#>   featureNames: 3-methyl-2-oxovalerate 5'-methylthioadenosine ...
#>     sedoheptulose-7-phosphate (1063 total)
#>   fvarLabels: feature_ID dataset ... formula (13 total)
#>   fvarMetadata: labelDescription
#> experimentData: use 'experimentData(object)'
#> Annotation:  
#> - - - Processing information - - -
#>  MSnbase version: 2.24.2 
exprs(METAB_MSNSET)[1:10, 1:5] # assayData (first 10 rows and 5 columns)
#>                        90245014701 90248014701  90252014701 90265014701
#> 3-methyl-2-oxovalerate    1.178894   0.9060579  0.008567142  -0.8926003
#> 5'-methylthioadenosine   -3.528231  -4.2752777 -7.236768016  -6.3275334
#> adenosine                 4.437239   4.1397055  4.679195002   4.1583898
#> ADP                       3.410435   3.0386412  4.437419080   3.9750911
#> Alanine                         NA  -0.4870398 -0.179877286  -0.4590593
#> AMP                       2.528949   0.9225197  2.897239407   2.4056311
#> Anserine                        NA  -7.1303497 -0.209715023  -1.5475741
#> Arginine                        NA  -2.3855330 -2.218783654  -2.5437574
#> Asparagine                      NA  -2.4311125 -1.990426951  -2.3243017
#> Aspartic Acid                   NA  -2.2169214 -2.077245870  -1.5561938
#>                        90266014701
#> 3-methyl-2-oxovalerate   0.4339056
#> 5'-methylthioadenosine  -4.1024668
#> adenosine                4.5390403
#> ADP                      3.9017748
#> Alanine                 -0.3163109
#> AMP                      2.6180835
#> Anserine                -2.5877989
#> Arginine                -1.6949176
#> Asparagine              -1.8801836
#> Aspartic Acid           -1.6443759
head(fData(METAB_MSNSET))      # featureData
#>                                    feature_ID          dataset
#> 3-methyl-2-oxovalerate 3-methyl-2-oxovalerate       metab-t-ka
#> 5'-methylthioadenosine 5'-methylthioadenosine      metab-t-nuc
#> adenosine                           adenosine metab-u-hilicpos
#> ADP                                       ADP      metab-t-nuc
#> Alanine                               Alanine   metab-t-amines
#> AMP                                       AMP      metab-t-nuc
#>                                   name_in_figures refmet_super_class
#> 3-methyl-2-oxovalerate 3-Methyl-2-oxovaleric acid      Organic acids
#> 5'-methylthioadenosine     5'-Methylthioadenosine      Nucleic acids
#> adenosine                               Adenosine      Nucleic acids
#> ADP                                           ADP      Nucleic acids
#> Alanine                                   Alanine      Organic acids
#> AMP                                           AMP      Nucleic acids
#>                               refmet_main_class       refmet_sub_class
#> 3-methyl-2-oxovalerate               Keto acids Short-chain keto acids
#> 5'-methylthioadenosine                  Purines Purine ribonucleosides
#> adenosine                               Purines Purine ribonucleosides
#> ADP                                     Purines            Purine rNDP
#> Alanine                Amino acids and peptides            Amino acids
#> AMP                                     Purines            Purine rNMP
#>                        lipid_class chain_length double_bond    rt       mz
#> 3-methyl-2-oxovalerate        <NA>           NA          NA  0.00   0.0000
#> 5'-methylthioadenosine        <NA>           NA          NA  0.00   0.0000
#> adenosine                     <NA>           NA          NA  5.03 268.1034
#> ADP                           <NA>           NA          NA  0.00   0.0000
#> Alanine                       <NA>           NA          NA 13.30 260.3000
#> AMP                           <NA>           NA          NA  0.00   0.0000
#>                        neutral_mass       formula
#> 3-methyl-2-oxovalerate     130.0630       C6H10O3
#> 5'-methylthioadenosine     297.0896   C11H15N5O3S
#> adenosine                  267.0968    C10H13N5O4
#> ADP                        427.0294 C10H15N5O10P2
#> Alanine                     89.0477       C3H7NO2
#> AMP                        347.0631   C10H14N5O7P
head(pData(METAB_MSNSET))      # phenoData
#>                  pid   bid       labelid    sex timepoint exp_group
#> 90245014701 10043527 90245 4.458696e-313 Female       SED     F_SED
#> 90248014701 10043950 90248 4.458844e-313 Female       SED     F_SED
#> 90252014701 10044337 90252 4.459042e-313 Female       SED     F_SED
#> 90265014701 10046119 90265 4.459684e-313 Female       SED     F_SED
#> 90266014701 10046380 90266 4.459734e-313 Female       SED     F_SED
#> 90560014701 10700102 90560 4.474259e-313 Female        1W      F_1W