Mitochondrial DNA qPCR

Mitochondrial DNA quantitative polymerase chain reaction (qPCR) data.

MITO_DNA

Format

data.frame with 30 rows and 8 columns:

bid

integer; unique 5 digit identifier of all samples collected for an acute test/sample collection period. All samples collected during that period will have the same BID.

sex

factor; the sex of the rat with 2 levels "Female" and "Male".

timepoint

factor; exercise training group. Either "SED" (sedentary) or the number of weeks of training ("1W", "2W", "4W", "8W").

exp_group

factor; unique combination of sex (first letter) and timepoint.

mean_delta_CT

numeric; mean of the duplicate \(\Delta C_T\) values from each sample.

SE_delta_CT

numeric; standard error of mean_delta_CT.

delta_delta_CT

numeric; mean_delta_CT values centered on the mean of the "F_SED" exp_group.

relative_expr

numeric; \(2^{-\Delta \Delta C_T}\) values.

Details

Quantification of mitochondrial DNA (mtDNA) was performed and described by Amar, et al. (https://doi.org/10.1101/2023.01.13.523698). Briefly, real-time quantitative PCR was performed in duplicate for each of the scWAT samples selected for -omics analysis. The \(2^{-\Delta \Delta C_T}\) method (https://doi.org/10.1006/meth.2001.1262) was then applied to estimate the relative expression of the mitochondrial D-loop. Since both target (D-loop) and internal control (\(\beta\)-actin) were amplified in the same well, \(\Delta C_T\) was calculated as the mean of (\(C_{T,\beta-loop} - C_{T,\beta-actin}\)) for each sample. Then, \(\Delta \Delta C_T\) values were obtained by subtracting each \(\Delta C_T\) value by the mean \(\Delta C_T\) of the sedentary female group (the calibrator).

References

Amar, D., Gay, N. R., Jimenez-Morales, D., Beltran, P. M. J., Ramaker, M. E., Raja, A. N., Zhao, B., Sun, Y., Marwaha, S., Gaul, D., Hershman, S. G., Xia, A., Lanza, I., Fernandez, F. M., Montgomery, S. B., Hevener, A. L., Ashley, E. A., Walsh, M. J., Sparks, L. M., … The MoTrPAC Study Group (2023). The mitochondrial multi-omic response to exercise training across tissues. BioRxiv. https://doi.org/10.1101/2023.01.13.523698

Livak, K. J., & Schmittgen, T. D. (2001). Analysis of Relative Gene Expression Data Using Real-Time Quantitative PCR and the 2−ΔΔCT Method. Methods, 25(4), 402–408. https://doi.org/10.1006/meth.2001.1262

Examples

head(MITO_DNA, 10)
#> # A tibble: 10 × 8
#>      bid sex    timepoint exp_group mean_delta_CT SE_delta_CT delta_delta_CT
#>    <dbl> <fct>  <fct>     <fct>             <dbl>       <dbl>          <dbl>
#>  1 90245 Female SED       F_SED             -6.10     0.127          0.181  
#>  2 90248 Female SED       F_SED             -6.48     0.0819        -0.207  
#>  3 90252 Female SED       F_SED             -6.28     0.133         -0.00769
#>  4 90265 Female SED       F_SED             -6.82     0.103         -0.540  
#>  5 90266 Female SED       F_SED             -5.70     0.00273        0.574  
#>  6 90559 Female 1W        F_1W              -5.96     0.0392         0.320  
#>  7 90560 Female 1W        F_1W              -6.03     0.145          0.251  
#>  8 90564 Female 1W        F_1W              -5.72     0.0179         0.561  
#>  9 90567 Female 1W        F_1W              -6.36     0.0705        -0.0864 
#> 10 90571 Female 1W        F_1W              -6.99     0.103         -0.710  
#> # ℹ 1 more variable: relative_expr <dbl>