MotrpacRatTraining6moWATData: Data and Analysis Results for the MoTrPAC PASS1B Rat White Adipose Tissue Companion Paper

Contains data and code necessary to reproduce analyses that appear in "Sexual dimorphism and the multi-omic response to exercise training in rat subcutaneous white adipose tissue".

See also

Author

Maintainer: Tyler J. Sagendorf tyler.sagendorf@pnnl.gov (ORCID)

Authors:

Zhenxin Hou (ORCID) (Provided WGCNA-related code, triacylglycerol abundance data, and both notes and data regarding the redundant metabolite filtering process.)

Other contributors:

Christopher B. Newgard (Provided clinical analyte data.) [data contributor]

Michael J. Muehlbauer (Provided clinical analyte data.) [data contributor]

Pasquale Nigro (ORCID) (Provided histology data.) [data contributor]

Malene E. Lindholm (ORCID) (Provided mtDNA qPCR results.) [data contributor]