Plasma clinical analytes

A set of 9 common clinical analytes measured in plasma.

Usage

PLASMA_ANALYTES

Format

An object of class data.frame with 146 rows and 21 columns.

plate

integer; 96-well plate identifier (1--4).

runseq

integer; run order (1--146).

platepos

integer; position in each 96-well plate.

pid

integer; randomly generated 8-digit identifier used in linkage to phenotypic data. All samples from the same animal have the same PID.

bid

integer; unique 5 digit identifier of all samples collected for an acute test/sample collection period. All samples collected during that period will have the same BID.

labelid

integer; unique 11 digit specimen label identifier, originating at the collection site, that provides a link to specimen processing and is used for shipments to the biorepository. Same as viallabel only in instances where aliquots are not further processed at the biorepository.

omics_subset

logical; whether the sample was selected for -omics analysis.

sex

factor; the sex of the rat with levels "Female" and "Male".

timepoint

factor; exercise training group. Either "SED" (sedentary) or the number of weeks of training ("1W", "2W", "4W", "8W").

total_ketones

numeric; total ketone bodies (\(\mu\)mol/L).

nefa

numeric; non-esterified fatty acids (NEFA) (mmol/L).

lactate

numeric; lactate (mmol/L).

glucose

numeric; glucose (mg/dL).

glycerol

numeric; glycerol (mg/dL).

glucagon

numeric; glucagon (pmol).

insulin

numeric; insulin (pg/mL).

insulin_iu

numeric; insulin (\(\mu\)IU/mL) using the conversion 1 pg/mL = 0.023 \(\mu\)IU/mL provided in the package insert.

leptin

numeric; leptin (pg/mL).

corticosterone

numeric; corticosterone (ng/mL).

ins_gcg_ratio

numeric; molar insulin/glucagon ratio. Calculated as (insulin (pg/mL) / 5.804 (Da)) / glucagon (pmol).

homa_ir

numeric; Homeostatic Model Assessment for Insulin Resistance (HOMA-IR). Insulin (\(\mu\)IU/mL) x glucose (mg/dL) / 405.

Details

Fujifilm Wako (Osaka, Japan):

Total ketones: 415-73301, 411-73401

NEFA: 995-34791, 999-34691, 993-35191, 991-34891

Beckman (Brea, CA):

Lactate: A95550

Glucose: B24985

Glycerol: 445850

Meso Scale Discovery (Rockville, MD):

Glucagon: K1535YK

Insulin, Leptin: K15158C (multiplex assay)

Alpco (Salem, NH):

Corticosterone: 55-CORMS-E01