Transcriptomics MSnSet

An object of class MSnbase::MSnSet containing transcriptomics data for differential analysis and (once transformed to log\(_2\) counts per million reads) WGCNA.

Usage

TRNSCRPT_MSNSET

Format

Object of class MSnbase::MSnSet with 16404 features and 48 samples.

assayData

Object of class matrix with 16404 rows and 48 columns containing RSEM gene counts.

featureData

Object of class data.frame with 16404 rows and 3 variables containing additional feature information:

feature_ID

character; Reference Sequence (RefSeq) protein identifier.

gene_symbol

character; official gene symbol.

entrez_gene

integer; Entrez gene identifier.

phenoData

Object of class data.frame with 48 rows and 13 variables containing additional sample information:

pid

integer; randomly generated 8-digit identifier used in linkage to phenotypic data. All samples from the same animal have the same PID.

bid

integer; unique 5 digit identifier of all samples collected for an acute test/sample collection period. All samples collected during that period will have the same BID.

labelid

integer; unique 11 digit specimen label identifier, originating at the collection site, that provides a link to specimen processing and is used for shipments to the biorepository. Same as viallabel only in instances where aliquots are not further processed at the biorepository.

viallabel

character; unique 11 digit sample vial identifier. Starts with the bid.

sex

factor; the sex of the rat with levels "Female" and "Male".

timepoint

factor; exercise training group. Either "SED" (sedentary) or the number of weeks of training ("1W", "2W", "4W", "8W").

exp_group

factor; experimental group. Unique combination of sex (first letter) and timepoint.

rin

numeric; RNA integrity number (RIN).

pct_globin

numeric; percent of reads mapping to globin.

pct_umi_dup

numeric; percent of PCR duplicates as quantified with Unique Molecular Identifiers (UMIs).

median_5_3_bias

numeric; median 5'-3' bias.

Details

The original 32883 transcripts were filtered with filterByExpr using group = TRNSCRPT_MSNSET[["exp_group"]]. Then, library sizes were TMM normalized with calcNormFactors. Two outlier samples (90423017005 and 90410017005) were removed before variables pct_globin, rin, pct_umi_dup, and median_5_3_bias were mean-imputed, centered, and scaled.

OUTLIERS for sample outlier information.

FEATURE_TO_GENE for details of feature-to-gene mapping.

See also

Examples

library(MSnbase)

TRNSCRPT_MSNSET                   # summary
#> MSnSet (storageMode: lockedEnvironment)
#> assayData: 16404 features, 48 samples 
#>   element names: exprs 
#> protocolData: none
#> phenoData
#>   sampleNames: 90245017005 90248017005 ... 90227017005 (48 total)
#>   varLabels: pid bid ... median_5_3_bias (11 total)
#>   varMetadata: labelDescription
#> featureData
#>   featureNames: ENSRNOG00000000001 ENSRNOG00000000007 ...
#>     ENSRNOG00000062306 (16404 total)
#>   fvarLabels: feature_ID gene_symbol entrez_gene
#>   fvarMetadata: labelDescription
#> experimentData: use 'experimentData(object)'
#> Annotation:  
#> - - - Processing information - - -
#>  MSnbase version: 2.24.2 

exprs(TRNSCRPT_MSNSET)[1:10, 1:5] # assayData (first 10 rows and 5 columns)
#>                    90245017005 90248017005 90252017005 90265017005 90266017005
#> ENSRNOG00000000001          28          14          23          28          35
#> ENSRNOG00000000007           5          12           7           7           1
#> ENSRNOG00000000008         110          53         114          88          87
#> ENSRNOG00000000010          54          17           3           7          14
#> ENSRNOG00000000012         106         153         158         103         209
#> ENSRNOG00000000017          34          54          61          37          64
#> ENSRNOG00000000021         721         321         326         367         436
#> ENSRNOG00000000024         493         773         674         644         847
#> ENSRNOG00000000033         512         323         403         377         411
#> ENSRNOG00000000034         490         279         425         301         447

head(fData(TRNSCRPT_MSNSET))      # featureData
#>                            feature_ID    gene_symbol entrez_gene
#> ENSRNOG00000000001 ENSRNOG00000000001 AABR07013255.1        <NA>
#> ENSRNOG00000000007 ENSRNOG00000000007           Gad1       24379
#> ENSRNOG00000000008 ENSRNOG00000000008           Alx4      296511
#> ENSRNOG00000000010 ENSRNOG00000000010          Cbln1      498922
#> ENSRNOG00000000012 ENSRNOG00000000012          Tcf15      296272
#> ENSRNOG00000000017 ENSRNOG00000000017         Steap1      297738

head(pData(TRNSCRPT_MSNSET))      # phenoData
#>                  pid   bid       labelid   viallabel    sex timepoint exp_group
#> 90245017005 10043527 90245 4.458696e-313 90245017005 Female       SED     F_SED
#> 90248017005 10043950 90248 4.458844e-313 90248017005 Female       SED     F_SED
#> 90252017005 10044337 90252 4.459042e-313 90252017005 Female       SED     F_SED
#> 90265017005 10046119 90265 4.459684e-313 90265017005 Female       SED     F_SED
#> 90266017005 10046380 90266 4.459734e-313 90266017005 Female       SED     F_SED
#> 90560017005 10700102 90560 4.474259e-313 90560017005 Female        1W      F_1W
#>                    rin pct_globin pct_umi_dup median_5_3_bias
#> 90245017005  0.1534937 -0.3394055   3.1014362      0.98444319
#> 90248017005 -2.0568155  1.0182164   0.3599009     -2.07005015
#> 90252017005 -2.5479953  0.3394055  -0.0376748     -1.03230562
#> 90265017005 -1.5656357 -0.3394055  -0.5626291      0.53410122
#> 90266017005 -0.3376861  0.3394055  -0.4034058     -0.05982804
#> 90560017005 -2.3024054  1.6970274  -0.2760272     -1.97214972