Differential analysis results

Differential analysis of proteins, transcripts, phosphorylation sites, and metabolites measured in rat scWAT.

PROT_DA     # proteomics
PHOSPHO_DA  # phosphoproteomics
TRNSCRPT_DA # transcriptomics
METAB_DA    # metabolomics

Format

A named list of 3 data.frame objects:

"trained_vs_SED"

All trained timepoints compared to their sex-matched sedentary control group. Example: F_1W - F_SED. Total of 8 contrasts.

"MvF_SED"

Sedentary males compared to sedentary females. Only 1 contrast.

"MvF_exercise_response"

Comparisons of the male and female training responses. Example: (M_8W - M_SED) - (F_8W - F_SED). Total of 4 contrasts.

In addition to all columns present in the Biobase::fData tables, each data.frame contains the following 8 variables:

logFC

numeric; difference in the weighted mean log\(_2\) values of the groups specified in the contrast column.

AveExpr

numeric; mean log\(_2\) relative abundance of each feature.

t

numeric; LIMMA moderated t-statistic.

P.Value

numeric; p-value.

adj.P.Val

numeric; Benjamini-Hochberg adjusted p-value. P-values are adjusted across all contrasts.

feature

character; uniquely defines each feature that was tested. Same as feature_ID.

contrast

factor; the contrast(s) being tested.

se

numeric; standard error (SE) of the logFC column.

Details

Differential analysis was performed with MotrpacRatTraining6moWAT::limma_full on the ExpressionSet objects.

See also

Examples

## Number of differential features (-ome-wide FDR < 0.05)
# Convenience function
f1 <- function(x) {
  lapply(x, function(.x) with(.x, table(contrast, adj.P.Val < 0.05)))
}

f1(PROT_DA)     # Proteins
#> $trained_vs_SED
#>               
#> contrast       FALSE TRUE
#>   F_1W - F_SED  9875   89
#>   F_2W - F_SED  9743  221
#>   F_4W - F_SED  9825  139
#>   F_8W - F_SED  9839  125
#>   M_1W - M_SED  9811  153
#>   M_2W - M_SED  9888   76
#>   M_4W - M_SED  9393  571
#>   M_8W - M_SED  9310  654
#> 
#> $MvF_SED
#>                
#> contrast        FALSE TRUE
#>   M_SED - F_SED  5738 4226
#> 
#> $MvF_exercise_response
#>                                  
#> contrast                          FALSE TRUE
#>   (M_1W - M_SED) - (F_1W - F_SED)  9943   21
#>   (M_2W - M_SED) - (F_2W - F_SED)  9926   38
#>   (M_4W - M_SED) - (F_4W - F_SED)  9824  140
#>   (M_8W - M_SED) - (F_8W - F_SED)  9943   21
#> 
f1(PHOSPHO_DA)  # Phosphosites
#> $trained_vs_SED
#>               
#> contrast       FALSE  TRUE
#>   F_1W - F_SED 30065   239
#>   F_2W - F_SED 29952   352
#>   F_4W - F_SED 30083   221
#>   F_8W - F_SED 30147   157
#>   M_1W - M_SED 30235    69
#>   M_2W - M_SED 30166   138
#>   M_4W - M_SED 30022   282
#>   M_8W - M_SED 29360   944
#> 
#> $MvF_SED
#>                
#> contrast        FALSE  TRUE
#>   M_SED - F_SED 24276  6028
#> 
#> $MvF_exercise_response
#>                                  
#> contrast                          FALSE
#>   (M_1W - M_SED) - (F_1W - F_SED) 30304
#>   (M_2W - M_SED) - (F_2W - F_SED) 30304
#>   (M_4W - M_SED) - (F_4W - F_SED) 30304
#>   (M_8W - M_SED) - (F_8W - F_SED) 30304
#> 
f1(TRNSCRPT_DA) # Transcripts
#> $trained_vs_SED
#>               
#> contrast       FALSE  TRUE
#>   F_1W - F_SED 16237   167
#>   F_2W - F_SED 16345    59
#>   F_4W - F_SED 16349    55
#>   F_8W - F_SED 16211   193
#>   M_1W - M_SED 16248   156
#>   M_2W - M_SED 15875   529
#>   M_4W - M_SED 16391    13
#>   M_8W - M_SED 16314    90
#> 
#> $MvF_SED
#>                
#> contrast        FALSE  TRUE
#>   M_SED - F_SED  6068 10336
#> 
#> $MvF_exercise_response
#>                                  
#> contrast                          FALSE  TRUE
#>   (M_1W - M_SED) - (F_1W - F_SED) 16270   134
#>   (M_2W - M_SED) - (F_2W - F_SED) 16314    90
#>   (M_4W - M_SED) - (F_4W - F_SED) 16390    14
#>   (M_8W - M_SED) - (F_8W - F_SED) 16361    43
#> 
f1(METAB_DA)    # Metabolites
#> $trained_vs_SED
#>               
#> contrast       FALSE TRUE
#>   F_1W - F_SED   969   94
#>   F_2W - F_SED   928  135
#>   F_4W - F_SED  1043   20
#>   F_8W - F_SED   924  139
#>   M_1W - M_SED   973   90
#>   M_2W - M_SED   902  161
#>   M_4W - M_SED   909  154
#>   M_8W - M_SED   852  211
#> 
#> $MvF_SED
#>                
#> contrast        FALSE TRUE
#>   M_SED - F_SED   448  615
#> 
#> $MvF_exercise_response
#>                                  
#> contrast                          FALSE TRUE
#>   (M_1W - M_SED) - (F_1W - F_SED)  1027   36
#>   (M_2W - M_SED) - (F_2W - F_SED)  1025   38
#>   (M_4W - M_SED) - (F_4W - F_SED)  1007   56
#>   (M_8W - M_SED) - (F_8W - F_SED)  1031   32
#>